Create Uncertainty analysis
The primary objective of this analysis was to quantify the level of uncertainty associated with the buffalo mozzarella cheese WF, with a focus on different Wfc (green, blue, and grey) and various sources of water use. Uncertainties were estimated by coupling a bootstrap technique with Monte Carlo (MC) simulations in a Python environment, where values within predefined ranges were randomly selected for each parameter . This stochastic process was repeated 10,000 times, thereby generating a comprehensive uncertainty distribution.